Remove unnecessary cast

+++ src/watch_linux.rs
@@ -59,10 +59,7 @@ fn start_watcher_thread<F: FnMut(Update) + Send + 'static>(
         Some(SockProtocol::NetlinkRoute),
     )
     .map_err(|e| Error::CreateSocket(e.to_string()))?;
-    let sa_nl = NetlinkAddr::new(
-        0,
-        (RTMGRP_LINK | RTMGRP_IPV4_IFADDR | RTMGRP_IPV6_IFADDR) as u32,
-    );
+    let sa_nl = NetlinkAddr::new(0, RTMGRP_LINK | RTMGRP_IPV4_IFADDR | RTMGRP_IPV6_IFADDR);
     bind(sockfd.as_raw_fd(), &sa_nl).map_err(|e| Error::Bind(e.to_string()))?;
     let (pipe_rd, pipe_wr) = pipe().map_err(|e| Error::CreatePipe(e.to_string()))?;
